Avoid jumping to address zero when calling pthread_atfork() on Linux.
On Linux, pthread_atfork() is provided not by libpthread.so.0 but by
libpthread_nonshared.a, and using "#pragma weak pthread_atfork" will
cause pthread_atfork to be undefined (and have an address of NULL)
even if libpthread_nonshared.a was explicitly specified to the linker.

To avoid calls to pthread_atfork() jumping to address zero on Linux,
avoid using "#pragma weak pthread_atfork" if libpthread_nonshared.a is
available.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>