Fix ->set_poll_timeout() related bug with timers that expire at time zero.
If the ivykis core decides to use ->set_poll_timeout() to offload the
handling of a timer, the epoll_create(2) / port_create(3C) poll
methods pass the given timeout directly into the timerfd_settime(2) /
timer_settime(3C) system calls, but if an ivykis user has registered
an iv_timer with an expiration time of (tv_sec, tv_nsec) = (0, 0),
the *_settime() system calls will interpret such a zero timeout value
as a request to disable the timer, and as a result, no timer handlers
will be run.

To work around this, we'll pass a timeout of (tv_sec, tv_nsec) = (0, 1)
into these system calls if the next expiring timer expires at (0, 0).

This bug was reported by and tracked down with the kind help of
Brian De Wolf (github: bldewolf), who ran into this issue with
syslog-ng 3.9.1.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>