Fix object reference counting bug when unregistering an expired timer.
If two timers expire during the same poll epoch, and the timer whose
handler happens to run first unregisters the other timer, we would
accidentally decrement the count of objects registered to the current
thread's ivykis event loop, st->numobjs, one too many times.

This could cause the object count to drop to zero, and iv_main() to
return while there are still objects registered, or it could cause the
object count to go negative, and iv_main() to not return even though
it should.

This patch fixes the logic in iv_timer_unregister() to only decrement
st->numobjs if the timer we are unregistering was on the timer heap,
and to avoid decrementing if the timer is on the list of expired
timers, and that fixes this bug.  And while we're at it, let's add a
test case to catch this bug.

Fixes: 7190532aca90 ("iv_timer: Implement fairness for timers.")
Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>