Explicitly build with -D_REENTRANT on Solaris.
Since commit 3c9bdf7a8bd2 ("Stop linking against libpthread.so by
default."), we stopped building with -pthreads on Solaris, and that
started causing this error at runtime when linked into a
multithreaded application:

	iv_fd_port_poll: got error 0[Error 0]

As László Várady explains:

	So the problem here is that after the -pthread C flag has
	been removed, ivykis started to operate on a non-threadsafe
	errno (on Solaris-like systems).  This is because -pthread
	implies -D_REENTRANT.

On Solaris, compiling a .c file without _REENTRANT and without
_POSIX_C_SOURCE set to some value >= 199506 will cause 'errno' to refer
to the main (first) thread's errno variable, instead of to the errno
variable for the thread we are running in.  From /usr/include/errno.h
on a Solaris 11 machine:

	#if defined(_REENTRANT) || defined(_TS_ERRNO) || _POSIX_C_SOURCE - 0 >= 199506L
	extern int *___errno();
	#define errno (*(___errno()))
	#else
	extern int errno;
	#endif

The reason for not unconditionally implementing errno with a function
call is presumably performance, as it will be slightly slower to access
errno by dereferencing the return value of a function call than by
accessing a bss variable directly, but for ivykis it's fine to always
go via the function call, as we don't access errno very often, and the
alternative would be to build ivykis twice, once with _REENTRANT and
once without, and then to deal with the complexity of picking the right
one to use, which sounds like it has potential for yet more screwups,
and which doesn't seem worth it.

Fixes: 3c9bdf7a8bd2 ("Stop linking against libpthread.so by default.")
Reported-by: Brian De Wolf (github: bldewolf)
Reported-by: Filip Hajny (github: mamash)
Tested-by: László Várady <laszlo.varady@balabit.com>
Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>