iv_work: add support for submitting work_items from within worker threads
This patch adds iv_work_pool_submit_continuation() function that allows the
submission of work items from any of the worker threads.

The idea is to make it possible for work item handlers to decide if some of
their tasks is best carried out by a separate work item, to be executed in
parallel.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>