iv_fd_epoll: Use the glibc-provided epoll_pwait2() wrapper.
When built for i386, the iv_fd_epoll poll method is getting -EINVAL
back from epoll_pwait2(2), presumably because the userland version and
the system call version of 'struct timespec' have different layouts,
and because we invoke the system call via the syscall() wrapper, which
means that no layout translation can be performed by glibc.

When this code was originally written, glibc did not provide an
epoll_pwait2(2) wrapper, and this is why we invoke it using syscall().
However, glibc does provide a wrapper now, and if we use that wrapper,
the issue seems to go away, so let's do that.

Issue originally reported by Peter Czanik:
https://github.com/buytenh/ivykis/issues/32

Reported-by: Peter Czanik <czanik@balabit.hu>
Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>
Tested-by: Balazs Scheidler <balazs.scheidler@axoflow.com>
Tested-by: László Böszörményi (GCS) <gcs@debian.org>